requirements.txt
django==1.7.8
djangorestframework==3.0.5
mapnik2==2.2.0
greenwich==0.3.0
